Move `commandToArgsArray` into legendary/library
