+++ src/backend/launcher.ts
@@ -63,7 +63,7 @@ import { gameManagerMap } from 'backend/storeManagers'
 import * as VDF from '@node-steam/vdf'
 import { readFileSync } from 'fs'
 import { LegendaryCommand } from './storeManagers/legendary/commands'
-import { Entries } from 'type-fest'
+import { commandToArgsArray } from './storeManagers/legendary/library'
 
 async function prepareLaunch(
   gameSettings: GameSettings,
@@ -977,63 +977,6 @@ async function getWinePath({
   return stdout.trim()
 }
 
-/**
- * Converts a LegendaryCommand to a parameter list passable to Legendary
- * @param command
- */
-function commandToArgsArray(command: LegendaryCommand): string[] {
-  const commandParts: string[] = []
-
-  if (command.subcommand) commandParts.push(command.subcommand)
-
-  // Some commands need special handling
-  switch (command.subcommand) {
-    case 'install':
-      commandParts.push(command.appName)
-      if (command.sdlList) {
-        commandParts.push('--install-tag=')
-        for (const sdlTag of command.sdlList)
-          commandParts.push('--install-tag', sdlTag)
-      }
-      break
-    case 'launch':
-      commandParts.push(command.appName)
-      if (command.extraArguments)
-        commandParts.push(...shlex.split(command.extraArguments))
-      break
-    case 'info':
-    case 'sync-saves':
-    case 'uninstall':
-      commandParts.push(command.appName)
-      break
-    case 'move':
-      commandParts.push(command.appName, command.newBasePath)
-      break
-    case 'eos-overlay':
-      commandParts.push(command.action)
-      break
-    case 'import':
-      commandParts.push(command.appName, command.installationDirectory)
-      break
-  }
-
-  // Append parameters (anything starting with -)
-  for (const [parameter, value] of Object.entries(
-    command
-  ) as Entries<LegendaryCommand>) {
-    if (!parameter.startsWith('-')) continue
-    if (!value) continue
-    // Boolean values (specifically `true`) have to be handled differently
-    // Parameters that have a boolean type are just signified
-    // by the parameter being present, they don't have a value.
-    // Thus, we only add the key (parameter) here, instead of the key & value
-    if (value === true) commandParts.push(parameter)
-    else commandParts.push(parameter, value.toString())
-  }
-
-  return commandParts
-}
-
 export {
   prepareLaunch,
   launchCleanup,
@@ -1044,6 +987,5 @@ export {
   runWineCommand,
   callRunner,
   getRunnerCallWithoutCredentials,
-  getWinePath,
-  commandToArgsArray
+  getWinePath
 }
