+++ src/backend/storeManagers/legendary/library.ts
@@ -47,7 +47,7 @@ import {
   installStore,
   libraryStore
 } from './electronStores'
-import { callRunner, commandToArgsArray } from '../../launcher'
+import { callRunner } from '../../launcher'
 import { dirname, join } from 'path'
 import { isOnline } from 'backend/online_monitor'
 import { update } from './games'
@@ -56,6 +56,8 @@ import { app } from 'electron'
 import { copySync } from 'fs-extra'
 import { LegendaryCommand } from './commands'
 import { LegendaryAppName, LegendaryPlatform, Path } from './commands/base'
+import shlex from 'shlex'
+import { Entries } from 'type-fest'
 
 const allGames: Set<string> = new Set()
 let installedGames: Map<string, InstalledJsonMetadata> = new Map()
@@ -739,4 +741,61 @@ export async function getGameSdl(
     )
     return []
   }
+}
+
+/**
+ * Converts a LegendaryCommand to a parameter list passable to Legendary
+ * @param command
+ */
+export function commandToArgsArray(command: LegendaryCommand): string[] {
+  const commandParts: string[] = []
+
+  if (command.subcommand) commandParts.push(command.subcommand)
+
+  // Some commands need special handling
+  switch (command.subcommand) {
+    case 'install':
+      commandParts.push(command.appName)
+      if (command.sdlList) {
+        commandParts.push('--install-tag=')
+        for (const sdlTag of command.sdlList)
+          commandParts.push('--install-tag', sdlTag)
+      }
+      break
+    case 'launch':
+      commandParts.push(command.appName)
+      if (command.extraArguments)
+        commandParts.push(...shlex.split(command.extraArguments))
+      break
+    case 'info':
+    case 'sync-saves':
+    case 'uninstall':
+      commandParts.push(command.appName)
+      break
+    case 'move':
+      commandParts.push(command.appName, command.newBasePath)
+      break
+    case 'eos-overlay':
+      commandParts.push(command.action)
+      break
+    case 'import':
+      commandParts.push(command.appName, command.installationDirectory)
+      break
+  }
+
+  // Append parameters (anything starting with -)
+  for (const [parameter, value] of Object.entries(
+    command
+  ) as Entries<LegendaryCommand>) {
+    if (!parameter.startsWith('-')) continue
+    if (!value) continue
+    // Boolean values (specifically `true`) have to be handled differently
+    // Parameters that have a boolean type are just signified
+    // by the parameter being present, they don't have a value.
+    // Thus, we only add the key (parameter) here, instead of the key & value
+    if (value === true) commandParts.push(parameter)
+    else commandParts.push(parameter, value.toString())
+  }
+
+  return commandParts
 }
